Extracts styles
