+++ packages/combobox/src/ComboboxGroup/ComboboxGroup.styles.ts
@@ -1,0 +1,34 @@
+import { css } from '@leafygreen-ui/emotion';
+import { Theme } from '@leafygreen-ui/lib';
+import { palette } from '@leafygreen-ui/palette';
+
+export const comboboxGroupStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    padding-top: 8px;
+  `,
+  [Theme.Dark]: css`
+    padding-top: 8px;
+  `,
+};
+
+export const comboboxGroupLabel = css`
+  cursor: default;
+  width: 100%;
+  padding: 0 12px 2px;
+  outline: none;
+  overflow-wrap: anywhere;
+  font-size: 12px;
+  line-height: 16px;
+  font-weight: bold;
+  text-transform: uppercase;
+  letter-spacing: 0.4px;
+`;
+
+export const comboboxGroupLabelThemeStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    color: ${palette.gray.dark1};
+  `,
+  [Theme.Dark]: css`
+    color: ${palette.gray.light1};
+  `,
+};
